+++ src/main/java/nl/ramsolutions/sw/magik/jacoco/Main.java
@@ -2,6 +2,7 @@ package nl.ramsolutions.sw.magik.jacoco;
 
 import nl.ramsolutions.sw.magik.jacoco.generators.HtmlReportGenerator;
 import nl.ramsolutions.sw.magik.jacoco.generators.JacocoXmlReportGenerator;
+import nl.ramsolutions.sw.magik.jacoco.generators.SonarXmlReportGenerator;
 import org.apache.commons.cli.CommandLine;
 import org.apache.commons.cli.CommandLineParser;
 import org.apache.commons.cli.DefaultParser;
@@ -66,6 +67,12 @@ public final class Main {
         .hasArg()
         .type(PatternOptionBuilder.FILE_VALUE)
         .build();
+    private static final Option OPTION_SONAR_XML = Option.builder()
+        .longOpt("sonar-xml")
+        .desc("Output Sonar XML report to file")
+        .hasArg()
+        .type(PatternOptionBuilder.FILE_VALUE)
+        .build();
     private static final Option OPTION_BUNDLE_NAME = Option.builder()
         .longOpt("bundle-name")
         .desc("Name of the bundle, defaults to 'Smallworld product'")
@@ -82,6 +89,7 @@ public final class Main {
         OPTIONS.addOption(OPTION_DISCARD_EXECUTABLE);
         OPTIONS.addOption(OPTION_HTML);
         OPTIONS.addOption(OPTION_JACOCO_XML);
+        OPTIONS.addOption(OPTION_SONAR_XML);
         OPTIONS.addOption(OPTION_BUNDLE_NAME);
     }
 
@@ -101,7 +109,8 @@ public final class Main {
     private static boolean showHelp(final CommandLine commandLine) {
         return commandLine.hasOption(OPTION_HELP)
             || !commandLine.hasOption(OPTION_HTML)
-               && !commandLine.hasOption(OPTION_JACOCO_XML);
+               && !commandLine.hasOption(OPTION_JACOCO_XML)
+               && !commandLine.hasOption(OPTION_SONAR_XML);
     }
 
     /**
@@ -165,6 +174,16 @@ public final class Main {
                 discardExecutable,
                 bundleName);
             xmlReportGenerator.run();
+        } else if (commandLine.hasOption(OPTION_SONAR_XML)) {
+            final File outputFile = (File) commandLine.getParsedOptionValue(OPTION_SONAR_XML);
+            final SonarXmlReportGenerator sonarXmlReportGenerator = new SonarXmlReportGenerator(
+                productPaths,
+                sourcePaths,
+                executionDataFile,
+                outputFile,
+                discardExecutable,
+                bundleName);
+            sonarXmlReportGenerator.run();
         }
     }