+++ CHANGES.md
@@ -4,6 +4,7 @@ Changes
 2.0.3 (unreleased)
 
 - Rename `--xml`-parameter to `--jacoco-xml`-parameter, and rename `XmlReportGenerator` to `JacocoXmlReportGenerator`.
+- Add `--sonar-xml` to generate the SonarQube "Generic test coverage report format".
 
 2.0.2 (2023-08-21)